Implemented a Velocity-Based Damage System in Unreal Engine Blueprints
I have implemented in the character blueprint a system that, based on a hit event, retrieves the velocity from the object hitting the character. It checks if the event is registered on the server and verifies if it is another character. If it is the server and not another character, it then checks if the velocity is above a minimum threshold. The velocity is then clamped between a maximum and minimum value. It is converted to damage by mapping the velocity from the maximum and minimum values to a maximum and minimum damage. Since we have yet to decide how we are handling damage, it currently just prints the damage on the screen.